+++ lib/utils/browser/index.ts
@@ -4,3 +4,20 @@ export {
   getCssVariableName,
   toCssVar,
 } from './css-utils.ts';
+
+export {
+  getDeviceInfo,
+  getMobileDeviceType,
+  isAndroid,
+  isBlackBerry,
+  isDesktop,
+  isIEMobile,
+  isIOS,
+  isIPad,
+  isIPhone,
+  isIPod,
+  isMobile,
+  isOperaMini,
+  isWPDesktop,
+} from './device-utils.ts';
+export type { MobileDevice } from './device-utils.ts';